Make telemetry optional
