+++ libraries/extensions/telemetry/tracing/Cargo.toml
@@ -6,10 +6,18 @@ license = "Apache-2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+telemetry = ["dep:opentelemetry", "dep:opentelemetry-jaeger"]
+
 [dependencies]
-opentelemetry = { version = "0.17", features = ["rt-tokio", "metrics"] }
-opentelemetry-jaeger = { version = "0.16", features = ["rt-tokio"] }
 tokio = { version = "1.24.2", features = ["full"] }
 tracing-subscriber = { version = "0.3.15", features = ["env-filter"] }
 eyre = "0.6.8"
 tracing = "0.1.36"
+opentelemetry = { version = "0.17", features = [
+    "rt-tokio",
+    "metrics",
+], optional = true }
+opentelemetry-jaeger = { version = "0.16", features = [
+    "rt-tokio",
+], optional = true }
